Add files via upload
StorySketch AI is a fun and easy project that turns your written stories or ideas into pictures using Artificial Intelligence. You just type a sentence like “a dragon flying over a rainbow,” and the AI creates a beautiful image of it in a few seconds. It uses a special AI model called Stable Diffusion to understand your words and draw them. You can run it on Google Colab or make a simple app using Streamlit. It’s perfect for kids, students, writers, and anyone who wants to see their imagination come to life as an image.StorySketch AI is a fun and easy project that turns your written stories or ideas into pictures using Artificial Intelligence. You just type a sentence like “a dragon flying over a rainbow,” and the AI creates a beautiful image of it in a few seconds. It uses a special AI model called Stable Diffusion to understand your words and draw them. You can run it on Google Colab or make a simple app using Streamlit. It’s perfect for kids, students, writers, and anyone who wants to see their imagination come to life as an image.